requirements.txt
albumentations
datasouls_antispoof
retinaface_pytorch
fastapi
python-multipart
uvicorn
